refactor: extract risk enforcer and cost tracker summary mixins
