+++ src/synthorg/budget/enforcer.py
@@ -7,8 +7,6 @@ as described in the Cost Controls section of the Operations design page.
 """
 
 import copy
-from datetime import UTC, datetime
-from functools import partial
 from types import MappingProxyType
 from typing import TYPE_CHECKING
 
@@ -28,14 +26,12 @@ from synthorg.budget.errors import (
     DailyLimitExceededError,
     ProjectBudgetExhaustedError,
     QuotaExhaustedError,
-    RiskBudgetExhaustedError,
 )
 from synthorg.budget.quota import (
     DegradationAction,
     DegradationConfig,
     always_allowed_result,
 )
-from synthorg.budget.risk_check import RiskCheckResult
 from synthorg.constants import BUDGET_ROUNDING_PRECISION
 from synthorg.notifications.dispatcher import NotificationDispatcher  # noqa: TC001
 from synthorg.observability import get_logger
@@ -62,12 +58,7 @@ from synthorg.observability.events.quota import (
     QUOTA_CHECK_DENIED,
 )
 from synthorg.observability.events.risk_budget import (
-    RISK_BUDGET_DAILY_LIMIT_EXCEEDED,
     RISK_BUDGET_ENFORCEMENT_CHECK,
-    RISK_BUDGET_LIMIT_EXCEEDED,
-    RISK_BUDGET_RECORD_ADDED,
-    RISK_BUDGET_RECORD_FAILED,
-    RISK_BUDGET_TASK_LIMIT_EXCEEDED,
 )
 
 if TYPE_CHECKING:
@@ -80,7 +71,6 @@ if TYPE_CHECKING:
     )
     from synthorg.budget.quota import QuotaCheckResult
     from synthorg.budget.quota_tracker import QuotaTracker
-    from synthorg.budget.risk_record import RiskRecord
     from synthorg.budget.risk_tracker import RiskTracker
     from synthorg.budget.tracker import CostTracker
     from synthorg.core.agent import AgentIdentity
@@ -90,10 +80,12 @@ if TYPE_CHECKING:
     from synthorg.providers.routing.resolver import ModelResolver
     from synthorg.security.risk_scorer import RiskScorer
 
+from synthorg.budget.risk_enforcer import BudgetEnforcerRiskMixin
+
 logger = get_logger(__name__)
 
 
-class BudgetEnforcer:
+class BudgetEnforcer(BudgetEnforcerRiskMixin):
     """Budget enforcement: pre-flight, in-flight, and auto-downgrade.
 
     Concurrency-safe via CostTracker's asyncio.Lock.  Pre-flight
@@ -700,173 +692,6 @@ class BudgetEnforcer:
             project_id=project_id or None,
         )
 
-    # ── Risk budget enforcement ─────────────────────────────────
-
-    async def check_risk_budget(
-        self,
-        agent_id: str,
-        task_id: str,
-        action_type: str,
-    ) -> RiskCheckResult:
-        """Pre-flight risk budget check.
-
-        Checks per-task, per-agent daily, and total daily risk limits
-        including the projected risk of the pending action.
-
-        Pre-flight checks are best-effort under concurrency (TOCTOU).
-        See class docstring.
-
-        Raises:
-            RiskBudgetExhaustedError: When a risk limit is exceeded
-                and enforcement is active.
-        """
-        risk_cfg = self._budget_config.risk_budget
-        if not risk_cfg.enabled or self._risk_tracker is None:
-            return RiskCheckResult()
-
-        logger.debug(
-            RISK_BUDGET_ENFORCEMENT_CHECK,
-            agent_id=agent_id,
-            task_id=task_id,
-            action_type=action_type,
-        )
-
-        try:
-            # Score the pending action for projected enforcement.
-            projected = 0.0
-            if self._risk_scorer is not None:
-                projected = self._risk_scorer.score(action_type).risk_units
-
-            day_start = daily_period_start()
-            t = self._risk_tracker
-            checks = (
-                (
-                    risk_cfg.per_task_risk_limit,
-                    partial(t.get_task_risk, task_id),
-                    RISK_BUDGET_TASK_LIMIT_EXCEEDED,
-                    "Per-task",
-                ),
-                (
-                    risk_cfg.per_agent_daily_risk_limit,
-                    partial(t.get_agent_risk, agent_id, start=day_start),
-                    RISK_BUDGET_DAILY_LIMIT_EXCEEDED,
-                    "Per-agent daily",
-                ),
-                (
-                    risk_cfg.total_daily_risk_limit,
-                    partial(t.get_total_risk, start=day_start),
-                    RISK_BUDGET_LIMIT_EXCEEDED,
-                    "Total daily",
-                ),
-            )
-            for limit, get_risk, event, label in checks:
-                self._enforce_risk_limit(
-                    limit,
-                    await get_risk(),
-                    projected,
-                    event,
-                    label,
-                    agent_id,
-                    task_id,
-                )
-        except MemoryError, RecursionError:
-            raise
-        except RiskBudgetExhaustedError:
-            raise
-        except Exception:
-            logger.exception(
-                RISK_BUDGET_ENFORCEMENT_CHECK,
-                agent_id=agent_id,
-                task_id=task_id,
-                reason="risk_check_error",
-            )
-
-        return RiskCheckResult(risk_units=projected)
-
-    def _enforce_risk_limit(  # noqa: PLR0913
-        self,
-        limit: float,
-        current: float,
-        projected: float,
-        event: str,
-        label: str,
-        agent_id: str,
-        task_id: str,
-    ) -> None:
-        """Check a single risk limit and raise if exceeded."""
-        if limit <= 0:
-            return
-        total = current + projected
-        if total >= limit:
-            logger.warning(
-                event,
-                agent_id=agent_id,
-                task_id=task_id,
-                current=current,
-                projected=projected,
-                limit=limit,
-            )
-            msg = f"{label} risk limit exceeded: {total:.2f} >= {limit:.2f}"
-            raise RiskBudgetExhaustedError(
-                msg,
-                agent_id=agent_id,
-                task_id=task_id,
-                risk_units_used=total,
-                risk_limit=limit,
-            )
-
-    async def record_risk(
-        self,
-        agent_id: str,
-        task_id: str,
-        action_type: str,
-    ) -> RiskRecord | None:
-        """Score and record a risk entry for the given action.
-
-        Returns ``None`` when risk budgets are disabled, no tracker
-        is configured, or no scorer is available.
-        """
-        from synthorg.budget.risk_record import (  # noqa: PLC0415
-            RiskRecord as _RiskRecord,
-        )
-
-        risk_cfg = self._budget_config.risk_budget
-        if (
-            not risk_cfg.enabled
-            or self._risk_tracker is None
-            or self._risk_scorer is None
-        ):
-            return None
-
-        try:
-            score = self._risk_scorer.score(action_type)
-            record = _RiskRecord(
-                agent_id=agent_id,
-                task_id=task_id,
-                action_type=action_type,
-                risk_score=score,
-                risk_units=score.risk_units,
-                timestamp=datetime.now(UTC),
-            )
-            await self._risk_tracker.record(record)
-        except MemoryError, RecursionError:
-            raise
-        except Exception:
-            logger.exception(
-                RISK_BUDGET_RECORD_FAILED,
-                agent_id=agent_id,
-                action_type=action_type,
-            )
-            return None
-        logger.info(
-            RISK_BUDGET_RECORD_ADDED,
-            agent_id=agent_id,
-            task_id=task_id,
-            action_type=action_type,
-            risk_units=score.risk_units,
-        )
-        return record
-
     # ── Private helpers ──────────────────────────────────────────
 
     async def _compute_baselines_safe(
@@ -884,7 +709,7 @@ class BudgetEnforcer:
                 daily_limit,
                 agent_id,
             )
-        except MemoryError, RecursionError:  # builtin MemoryError (OOM)
+        except MemoryError, RecursionError:
             raise
         except Exception:
             logger.exception(
@@ -926,23 +751,7 @@ class BudgetEnforcer:
         *,
         error_event: str = BUDGET_PREFLIGHT_ERROR,
     ) -> float | None:
-        """Query project cost from durable aggregate or in-memory tracker.
-
-        Returns the total cost (rounded to
-        ``BUDGET_ROUNDING_PRECISION``), or ``None`` when both
-        sources fail (caller should skip enforcement on ``None``).
-
-        Args:
-            project_id: Project identifier.
-            error_event: Event constant to log on failure.  Allows
-                callers to preserve distinct monitoring semantics
-                (e.g. preflight vs baseline).
-
-        Raises:
-            MemoryError: Re-raised unconditionally.
-            RecursionError: Re-raised unconditionally.
-        """
-        # Try durable aggregate first.
+        """Query project cost from durable aggregate or in-memory tracker."""
         if self._project_cost_repo is not None:
             try:
                 aggregate = await self._project_cost_repo.get(
@@ -956,7 +765,6 @@ class BudgetEnforcer:
                     project_id=project_id,
                     reason="project_cost_aggregate_query_failed",
                 )
-                # Fall through to in-memory.
             else:
                 raw = aggregate.total_cost if aggregate else 0.0
                 cost = round(raw, BUDGET_ROUNDING_PRECISION)
@@ -968,7 +776,6 @@ class BudgetEnforcer:
                 )
                 return cost
 
-        # Fallback to in-memory tracker.
         try:
             cost = await self._cost_tracker.get_project_cost(
                 project_id,
